Sent a set of data from mocks to the “Favorites” page. Removed component "FavoriteEmpty", create "FavoriteEmpty" component and display this component
