+++ src/components/favoriteCard/FavoriteCard.tsx
@@ -1,23 +1,28 @@
 import React from 'react';
 import { CardProps } from '../card';
 
-export const FavoriteCard: React.FC<CardProps> = ({ratingStars, placeType, placeName, price, imageSrc, premium}) => (
+export const FavoriteCard: React.FC<CardProps> = ({offer}) => (
   <article className="favorites__card place-card">
-    <div className="place-card__mark">
-      <span>{premium}</span>
-    </div>
+    {offer.isPremium ? (
+      <div className="place-card__mark">
+        <span>Premium</span>
+      </div>
+    ) : ''}
     <div className="favorites__image-wrapper place-card__image-wrapper">
       <a href="#">
-        <img className="place-card__image" src={imageSrc} width="150" height="110" alt="Place image"/>
+        <img className="place-card__image" src={offer.previewImage} width="150" height="110" alt="Place image"/>
       </a>
     </div>
     <div className="favorites__card-info place-card__info">
       <div className="place-card__price-wrapper">
         <div className="place-card__price">
-          <b className="place-card__price-value">&euro;{price}</b>
+          <b className="place-card__price-value">&euro;{offer.price}</b>
           <span className="place-card__price-text">&#47;&nbsp;night</span>
         </div>
-        <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
+        <button
+          className={offer.isFavorite ? 'place-card__bookmark-button button place-card__bookmark-button--active' : 'place-card__bookmark-button button'}
+          type="button"
+        >
           <svg className="place-card__bookmark-icon" width="18" height="19">
             <use xlinkHref="#icon-bookmark"></use>
           </svg>
@@ -26,14 +31,14 @@ export const FavoriteCard: React.FC<CardProps> = ({ratingStars, placeType, place
       </div>
       <div className="place-card__rating rating">
         <div className="place-card__stars rating__stars">
-          <span style={{'width': ratingStars}}></span>
+          <span style={{'width': offer.rating}}></span>
           <span className="visually-hidden">Rating</span>
         </div>
       </div>
       <h2 className="place-card__name">
-        <a href="#">{placeName}</a>
+        <a href="#">{offer.title}</a>
       </h2>
-      <p className="place-card__type">{placeType}</p>
+      <p className="place-card__type">{offer.type}</p>
     </div>
   </article>
 );
