+++ src/components/favoriteEmpty/index.ts
